Fixed command chaining, cleaned up noob code
Added a new global game object which controls global game data.
Rather than assigning the next step in the chain as an argument, the
'doNext' property is changed. There is a single, always listening
event, which runs doNext any time the enter key is pressed.
Jquery is in use again, to simplify the code, and to take care of any
syncronisation problems we might eventually have in the future.
Added array match functions to clean up if statements.